+++ jakartaee/src/main/todo-app/src/utils/ToDoService.js
@@ -2,7 +2,7 @@ import axios from 'axios';
 
 const API_URL = 'http://localhost:8080/resources/todo';
 
-export const ToDoService = {
+const ToDoService = {
   getItems: (userId) => {
     return axios.get(`${API_URL}/${userId}`);
   },
@@ -19,3 +19,5 @@ export const ToDoService = {
     return axios.delete(`${API_URL}/${userId}/${itemId}`);
   }
 };
+
+export default ToDoService;
